requirements.txt
appdirs==1.4.4
appnope==0.1.0
ariadne==0.12.0
attrs
backcall==0.2.0
black==20.8b1
brotlipy==0.7.0
certifi==2020.6.20
cffi
chardet==3.0.4
cryptography
cycler==0.10.0
decorator==4.4.2
eventkit==0.8.6
Flask==1.1.2
Flask-SQLAlchemy
graphql-core==3.0.5
idna
importlib-metadata
iniconfig
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
kiwisolver
loguru==0.5.3
MarkupSafe==1.1.1
matplotlib
more-itertools
mypy-extensions==0.4.3
nest-asyncio==1.4.2
numpy
olefile==0.46
packaging==20.4
pandas
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow
pluggy==0.13.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser
Pygments==2.7.2
pyOpenSSL
pyparsing==2.4.7
PySocks==1.7.1
pytest
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.26
regex==2020.10.28
requests
SQLAlchemy
starlette==0.13.8
toml
toolz==0.11.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3
wcwidth==0.2.5
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp

mysqlclient
Faker
python-dotenv
pydantic_sqlalchemy
Flask-Migrate
beautifulsoup4
tqdm
psycopg2-binary
tcgplayer-api
gunicorn
